requirements.txt
# Deep Learning Framework
torch>=2.0.0  # Current: 2.1.0
torchvision>=0.15.0  # Current: 0.16.0
pytorch-lightning>=2.0.0  # Current: 2.1.0

# Data Processing
numpy>=1.21.0  # Current: 1.24.0
Pillow>=9.0.0  # Current: 10.0.0
opencv-python>=4.5.0  # Current: 4.8.0

# Visualization
matplotlib>=3.5.0  # Current: 3.7.0
wandb>=0.15.0  # Current: 0.15.0

# Utilities
PyYAML>=6.0  # Current: 6.0
scikit-learn>=1.0.0  # Current: 1.3.0
tqdm>=4.65.0  # Current: 4.65.0

certifi>=2021.10.8
cffi>=1.15.0
mkl-fft>=1.3.6,<1.3.15
mkl-random>=1.2.2
olefile>=0.46
pycparser>=2.20
six>=1.16.0
matplotlib>=3.3.0
